+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unchanged]
+
+### Added
+- Add retrofit information to the mapping set class [#252](https://github.com/IN-CORE/incore-services/issues/252)
+
 ## [1.23.0] - 2023-12-13
 
 ### Changed